pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "boto3>=1.37.11",
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "gunicorn>=23.0.0",
    "marshmallow>=3.26.1",
    "openai>=1.66.2",
    "pytest>=8.3.5",
    "python-dotenv>=1.0.1",
    "replicate>=1.0.4",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
